Fixed signed/unsigned warnings

+++ plugins/controlboard/src/ControlBoardDriverCurrent.cpp
@@ -29,7 +29,7 @@ namespace yarp {
 
         bool GazeboYarpControlBoardDriver::setRefCurrent(int j, double v)
         {
-            if (j >= 0 && j < (int)m_numberOfJoints) {
+            if (j >= 0 && static_cast<size_t>(j) < m_numberOfJoints) {
                 m_jntReferenceTorques[j] = v*m_kPWM[j];
                 return true;
             }
@@ -50,7 +50,7 @@ namespace yarp {
         bool GazeboYarpControlBoardDriver::setRefCurrents(const double* v)
         {
             if (!v) return false;
-            for (unsigned int j = 0; j < m_numberOfJoints; ++j) {
+            for (size_t j = 0; j < m_numberOfJoints; ++j) {
                 m_jntReferenceTorques[j] = v[j]*m_kPWM[j];
             }
             return true;
@@ -82,7 +82,7 @@ namespace yarp {
 
         bool GazeboYarpControlBoardDriver::getRefCurrent(int j, double *v)
         {
-            if (v && j >= 0 && j < (int)m_numberOfJoints) {
+            if (v && j >= 0 && static_cast<size_t>(j) < m_numberOfJoints) {
                 *v = m_jntReferenceTorques[j];
                 return true;
             }
@@ -92,7 +92,7 @@ namespace yarp {
         bool GazeboYarpControlBoardDriver::getRefCurrents(double *v)
         {
             if (!v) return false;
-            for (unsigned int j = 0; j < m_numberOfJoints; ++j) {
+            for (size_t j = 0; j < m_numberOfJoints; ++j) {
                 v[j] = m_jntReferenceTorques[j];
             }
             return true;
